structure adjustment
