+++ csrc/apis/c/executor.h
@@ -1,0 +1,51 @@
+//
+// Created by zhangli on 4/26/22.
+//
+
+#ifndef MMDEPLOY_CSRC_APIS_C_EXECUTOR_H_
+#define MMDEPLOY_CSRC_APIS_C_EXECUTOR_H_
+
+#include "apis/c/common.h"
+
+#if __cplusplus
+extern "C" {
+#endif
+
+typedef struct mmdeploy_value* mmdeploy_value_t;
+typedef mmdeploy_value_t (*mmdeploy_invocable_t)(mmdeploy_value_t, void*);
+
+struct mmdeploy_sender;
+struct mmdeploy_scheduler;
+
+typedef mmdeploy_sender (*mmdeploy_kleisli_t)(mmdeploy_value_t, void*);
+
+typedef struct mmdeploy_sender* mmdeploy_sender_t;
+typedef struct mmdeploy_scheduler* mmdeploy_scheduler_t;
+
+MMDEPLOY_API mmdeploy_scheduler_t mmdeploy_inline_scheduler();
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_just(mmdeploy_value_t value);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_schedule(mmdeploy_scheduler_t scheduler);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_transfer(mmdeploy_sender_t input,
+                                                          mmdeploy_scheduler_t scheduler);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_then(mmdeploy_sender_t input,
+                                                      mmdeploy_invocable_t fn, void* context);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_let_value(mmdeploy_sender_t input,
+                                                           mmdeploy_kleisli_t kleisli,
+                                                           void* context);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_split(mmdeploy_sender_t input);
+
+MMDEPLOY_API mmdeploy_sender_t mmdeploy_executor_when_all(mmdeploy_sender_t* inputs, int32_t n);
+
+MMDEPLOY_API mmdeploy_value_t mmdeploy_executor_sync_wait(mmdeploy_sender_t input);
+
+#if __cplusplus
+}
+#endif
+
+#endif  // MMDEPLOY_CSRC_APIS_C_EXECUTOR_H_
